LayoutTests/fast/security-labels/attr-propagation-expected.txt

This tests that label propagation works on DOM attributes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS yesTag.isOn(body.getAttribute(name)) is true
PASS yesTag.isOn(attr.value) is true
PASS noTag.isOn(attr.value) is false
PASS yesTag.isOn(body.getAttribute(name)) is true
PASS yesTag.isOn(attr.value) is true
PASS noTag.isOn(attr.value) is false
PASS attr.value is "barfoo"
PASS yesTag.isOn(body.getAttribute(name)) is true
PASS yesTag.isOn(attr.value) is true
PASS noTag.isOn(attr.value) is false
PASS tag1.isOn(document.documentElement.innerHTML) is true
PASS tag2.isOn(document.documentElement.innerHTML) is true
PASS successfullyParsed is true

TEST COMPLETE
